requirements.txt
google-auth-oauthlib
google-api-python-client
openai>=1.0.0
notion-client
python-dotenv 